+++ app/notify_client/organisations_api_client.py
@@ -5,6 +5,8 @@ from notifications_python_client.errors import HTTPError
 from app.extensions import redis_client
 from app.notify_client import NotifyAdminAPIClient, cache
 
+from flask import current_app
+
 
 class OrganisationsClient(NotifyAdminAPIClient):
 
@@ -53,11 +55,12 @@ class OrganisationsClient(NotifyAdminAPIClient):
     def update_organisation(self, org_id, cached_service_ids=None, **kwargs):
         api_response = self.post(url="/organisations/{}".format(org_id), data=kwargs)
 
-        if cached_service_ids:
-            redis_client.delete(*map('service-{}'.format, cached_service_ids))
+        if current_app.config['NOTIFY_ADMIN_API_CACHE_ENABLED']:
+            if cached_service_ids:
+                redis_client.delete(*map('service-{}'.format, cached_service_ids))
 
-        if 'name' in kwargs:
-            redis_client.delete(f'organisation-{org_id}-name')
+            if 'name' in kwargs:
+                redis_client.delete(f'organisation-{org_id}-name')
 
         return api_response
